refactor(packageparser): move loading and validating `package.json` file to a separate file
